feat: improved atom parser
