+++ src/helpers/RSSFeedHelpers.ts
@@ -4,6 +4,7 @@ import { loadRedditFeed, redditJsonFeedUrl } from './redditFeedHelpers'
 import { Debug } from './Debug'
 import { timeout } from './Utils'
 import { safeFetch } from './safeFetch'
+import { parseAtomFeed } from './atomFeedHelpers'
 
 // tslint:disable-next-line:no-var-requires
 const util = require('react-native-util')
@@ -111,7 +112,7 @@ export const loadRSSFeed = async (url: string, xml: string, startTime = 0, downl
     parser.addListener('error', (err: string) => {
         throw new Error(err)
     })
-    return await new Promise<RSSFeedWithMetrics>((resolve, reject) => {
+    const rssFeed = await new Promise<RSSFeedWithMetrics>((resolve, reject) => {
         parser.parseString(xml, (err: string, result: any) => {
             if (err) {
                 reject(err)
@@ -130,6 +131,7 @@ export const loadRSSFeed = async (url: string, xml: string, startTime = 0, downl
             resolve(feedWithMetrics)
         })
     })
+    return rssFeed
 }
 
 const parseFeed = (json: any) => {
@@ -138,46 +140,6 @@ const parseFeed = (json: any) => {
     } else if (json.rss) {
         return parseRSSFeed(json)
     }
-}
-
-const getEntryDate = (entry: any): string | null => {
-    if (entry.published != null) {
-        return entry.published[0]
-    }
-    if (entry.updated != null) {
-        return entry.updated[0]
-    }
-    return null
-}
-
-const parseAtomFeed = (json: any) => {
-    const feed = json.feed
-    const rss: any = { items: [] }
-
-    if (feed.title) {
-        rss.title = typeof feed.title[0] === 'string' ? feed.title[0] : feed.title[0]?._ ?? ''
-    }
-    if (feed.icon) {
-        rss.icon = feed.icon[0]
-    }
-    if (feed.link) {
-        rss.url = feed.link[0].href[0]
-    }
-
-    rss.items = feed.entry.map((entry: any) => {
-        const entryDate = getEntryDate(entry)
-        const item: RSSItem = {
-            title: entry.title ? entry.title[0] : '',
-            description: entry.content ? entry.content[0]._ : '',
-            created: entryDate ? Date.parse(entryDate) : Date.now(),
-            link: entry.link ? entry.link[0].href[0] : '',
-            url: entry.link ? entry.link[0].href[0] : '',
-            media: getAtomEntryMedia(entry),
-        }
-        return item
-    })
-
-    return rss
 }
 
 const parseRSSFeed = (json: any) => {
@@ -236,23 +198,4 @@ const parseRSSFeed = (json: any) => {
         })
     }
     return rss
-}
-
-const getAtomEntryMedia = (entry: any): RSSMedia | undefined => {
-    const atomMediaGroup = entry['media:group']
-    const atomMediaThumbnail = atomMediaGroup?.[0]?.['media:thumbnail']?.[0]
-    if (atomMediaThumbnail != null) {
-        try {
-            return {
-                thumbnail: [{
-                    url: [atomMediaThumbnail.url[0]],
-                    width: [parseInt(atomMediaThumbnail.width[0], 10)],
-                    height: [parseInt(atomMediaThumbnail.height[0], 10)],
-                }],
-            }
-        } catch (e) {
-            return undefined
-        }
-    }
-    return undefined
 }
